powerpc: fix "partition-zero"/%BOOT booting for PowerPC targets.
This is a patch against r1272 which implements fully bootcode booting via
Apple boot partition descriptors. The current implementation in
libopenbios/bootcode_load.c hardcoded values for QUIK (the old
OldWorld bootloader), which is useless if you try to boot anything
else.

In particular, I implemented this to simplify my development and test
cycle on iQUIK (https://github.com/andreiw/quik), but I also tested
this with the NetBSD floppy. The later now boots to stage 2 as well,
although it tries using some functionality unavailable in the OpenBIOS
implementation when compared to Apple's.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>